Revert "refactor(app): improve folder structure add extension function in their respective files"
This reverts commit 3ee44fd93e6ade9d2b6d200a9f22a987aa9554d2.
