+++ app/src/main/java/com/my/kizzy/di/ServiceModule.kt
@@ -12,6 +12,7 @@
 
 package com.my.kizzy.di
 
+import com.android.girish.vlog.Vlog
 import com.my.kizzy.domain.repository.KizzyRepository
 import com.my.kizzy.rpc.KizzyRPC
 import com.my.kizzy.utils.Prefs
