--- app/src/main/java/com/my/kizzy/ui/screen/apps/AppsInfo.kt
@@ -1,19 +1,0 @@
-/*
- *
- *  ******************************************************************
- *  *  * Copyright (C) 2022
- *  *  * AppsInfo.kt is part of Kizzy
- *  *  *  and can not be copied and/or distributed without the express
- *  *  * permission of yzziK(Vaibhav)
- *  *  *****************************************************************
- *
- *
- */
-
-package com.my.kizzy.ui.screen.apps
-
-data class AppsInfo(
-    val name: String,
-    val pkg: String,
-    val isChecked: Boolean,
-)
